Add files via upload
Classification of emergency vehicles which has potential to save life of humans by adjusting the green signals and automating the paving ways for vehicles in self driving cars. So to overcome this problem we have built a Neural Network algorithm which is trained on a several thousands of images. With the help of keras we have created a model by taking required parameters (weights) then predicting the trained model on unseen data. To increase the accuracy of model we tuned the hyperparameters (Activation function, Epochs, Optimization function)